# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pet Manager API
  description: |
    API REST para la gestión de mascotas, paseos, citas y gastos.

    **Login con Google** (fuera de esta API REST, son navegaciones del navegador):
    1. El frontend navega a `GET http://localhost:8080/oauth2/authorization/google`.
    2. El backend redirige a Google y, al volver, a `{frontend}/auth/callback#token=<jwt>`
       (o a `{frontend}/login?error=google` si se cancela o falla).
    3. Ese token es un `bearerAuth` normal, igual que el de `/auth/login`.
  version: 1.3.1
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local (Spring Boot)

# Todas las operaciones exigen JWT salvo las que declaran `security: []` (registro y login)
security:
  - bearerAuth: []

tags:
  - name: Auth
    description: Registro, inicio de sesión y usuario actual
  - name: Pets
    description: Mascotas del usuario autenticado
  - name: Walks
    description: Paseos de una mascota
  - name: VetAppointments
    description: Citas veterinarias de una mascota
  - name: Expenses
    description: Gastos de una mascota

paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Crear una cuenta (devuelve ya un token de sesión)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Ya existe una cuenta con ese email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Iniciar sesión con email y contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Email o contraseña incorrectos (mismo error en ambos casos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Obtener el usuario de la sesión actual
      responses:
        '200':
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /pets:
    get:
      tags: [Pets]
      operationId: getPets
      summary: Obtener todas las mascotas del usuario autenticado
      responses:
        '200':
          description: Lista de mascotas ordenada por nombre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PetResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Pets]
      operationId: createPet
      summary: Añadir una nueva mascota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetRequest'
      responses:
        '201':
          description: Mascota creada exitosamente (sin cuerpo)
          headers:
            Location:
              description: URL de la mascota creada (/pets/{petId})
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /pets/{petId}:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Pets]
      operationId: getPet
      summary: Obtener el detalle de una mascota
      responses:
        '200':
          description: Mascota encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /pets/{petId}/photo:
    parameters:
      - $ref: '#/components/parameters/PetId'
    put:
      tags: [Pets]
      operationId: uploadPetPhoto
      summary: Sustituir la foto de una mascota
      description: >
        Sube la imagen a Supabase Storage, guarda su URL pública en `photoUrl` y borra la foto anterior.
        El formato se comprueba por el contenido del archivo, no por su extensión.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPG, PNG o WebP de hasta 5 MB
      responses:
        '200':
          description: Foto actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: La imagen supera los 5 MB
        '415':
          description: Formato no admitido (solo JPG, PNG o WebP)
        '502':
          description: No se pudo guardar la imagen en Supabase Storage

  /pets/{petId}/walks:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Walks]
      operationId: getPetWalks
      summary: Obtener los paseos de una mascota
      description: >
        Sin `date` devuelve todo el historial. Con `date` devuelve solo los paseos de ese día
        en la zona horaria `tz`. Orden: del más reciente al más antiguo.
      parameters:
        - name: date
          in: query
          required: false
          description: Día a consultar (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: tz
          in: query
          required: false
          description: Zona horaria IANA con la que se interpreta `date`
          schema:
            type: string
            default: Europe/Madrid
            example: Europe/Madrid
      responses:
        '200':
          description: Historial de paseos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WalkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Walks]
      operationId: createPetWalk
      summary: Registrar un paseo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkRequest'
      responses:
        '201':
          description: Paseo registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /pets/{petId}/vet-appointments:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [VetAppointments]
      operationId: getPetVetAppointments
      summary: Obtener las citas veterinarias de una mascota
      description: Ordenadas por fecha ascendente (pasadas y futuras).
      responses:
        '200':
          description: Lista de citas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VetAppointmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [VetAppointments]
      operationId: createPetVetAppointment
      summary: Registrar una cita veterinaria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VetAppointmentRequest'
      responses:
        '201':
          description: Cita registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetAppointmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /pets/{petId}/expenses:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Expenses]
      operationId: getPetExpenses
      summary: Obtener los gastos de una mascota
      description: Ordenados del más reciente al más antiguo.
      responses:
        '200':
          description: Lista de gastos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Expenses]
      operationId: createPetExpense
      summary: Registrar un gasto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '201':
          description: Gasto registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    PetId:
      name: petId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token devuelto por /auth/login o /auth/register, en la cabecera "Authorization: Bearer <token>"'

  responses:
    Unauthorized:
      description: Falta el token, no es válido o ha caducado
    BadRequest:
      description: Petición inválida (validación, JSON mal formado o parámetro incorrecto)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: La mascota no existe o pertenece a otro usuario
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    RegisterRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
        birthDate:
          type: string
          format: date
          description: Debe ser pasada

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    UserResponse:
      type: object
      required: [id, fullName, email, birthDate]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        birthDate:
          type: string
          format: date
          nullable: true

    AuthResponse:
      type: object
      required: [token, expiresAt, user]
      properties:
        token:
          type: string
          description: JWT firmado (HS256)
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserResponse'

    Species:
      type: string
      enum: [PERRO, GATO, OTRO]

    ExpenseCategory:
      type: string
      enum: [FOOD, VET, MEDICATION, GROOMING, TOYS, ACCESSORIES, INSURANCE, OTHER]

    PetResponse:
      type: object
      required: [id, name, species, breed, weightKg, personality, pathologies, photoUrl]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        species:
          $ref: '#/components/schemas/Species'
        breed:
          type: string
          nullable: true
        weightKg:
          type: number
          format: double
          nullable: true
        personality:
          type: string
          nullable: true
        pathologies:
          type: string
          nullable: true
          description: Lista separada por comas
        photoUrl:
          type: string
          nullable: true

    PetRequest:
      type: object
      required: [name, species]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        species:
          $ref: '#/components/schemas/Species'
        breed:
          type: string
          maxLength: 100
        weightKg:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 999.99
          multipleOf: 0.01
        personality:
          type: string
        pathologies:
          type: string
          description: Lista separada por comas

    WalkResponse:
      type: object
      required: [id, distanceKm, durationMinutes, didPee, didPoop, walkDatetime]
      properties:
        id:
          type: string
          format: uuid
        distanceKm:
          type: number
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        didPee:
          type: boolean
        didPoop:
          type: boolean
        walkDatetime:
          type: string
          format: date-time

    WalkRequest:
      type: object
      required: [didPee, didPoop, walkDatetime]
      properties:
        distanceKm:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 9999.99
          multipleOf: 0.01
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 1440
        didPee:
          type: boolean
        didPoop:
          type: boolean
        walkDatetime:
          type: string
          format: date-time
          description: No puede ser futura

    VetAppointmentResponse:
      type: object
      required: [id, description, cost, appointmentDate]
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        cost:
          type: number
          nullable: true
        appointmentDate:
          type: string
          format: date-time

    VetAppointmentRequest:
      type: object
      required: [appointmentDate]
      properties:
        description:
          type: string
          maxLength: 2000
        cost:
          type: number
          minimum: 0
          maximum: 99999999.99
          multipleOf: 0.01
        appointmentDate:
          type: string
          format: date-time
          description: Puede ser futura (citas próximas)

    ExpenseResponse:
      type: object
      required: [id, category, description, amount, expenseDate]
      properties:
        id:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        description:
          type: string
          nullable: true
        amount:
          type: number
        expenseDate:
          type: string
          format: date

    ExpenseRequest:
      type: object
      required: [category, amount, expenseDate]
      properties:
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        description:
          type: string
          maxLength: 2000
        amount:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 99999999.99
          multipleOf: 0.01
        expenseDate:
          type: string
          format: date
          description: No puede ser futura

    Error:
      type: object
      description: Formato de error por defecto de Spring Boot
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        path:
          type: string